Moved volume functions to separate file
Added volumes.go
